+++ src/fastapi_quickcrud/misc/schema_builder.py
@@ -371,8 +371,8 @@ class ApiParameterSchemaBuilder:
                 default = column.default.arg
             elif column.server_default is not None:
                 default = None
-            elif column.primary_key:
-                default = None
+            # elif column.primary_key:
+            #     default = None
             else:
                 default = ...
         else: